投稿後、フォームが空になるように変更する
