+++ src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 
-import postsReducer from "../components/features/posts/postlist/postsSlice";
+import postsReducer from "../components/features/posts/postsSlice";
 
 export const store = configureStore({
   reducer: {
