+++ src/components/features/posts/postlist/index.jsx
@@ -1,6 +1,6 @@
 import "./style.scss";
 import { useSelector, useDispatch } from "react-redux";
-import { postAllClear, postDeleted, postChecked } from "./postsSlice";
+import { postAllClear, postDeleted, postChecked } from "../postsSlice";
 
 export const PostList = () => {
   const dispatch = useDispatch();
